Create README.md
My app focuses on the stocks that is happening with the companies that I have chose is Apple, Tesla, Dell, Samsung, and Microsoft. We are going to include where a user buys an item from one of the companies and the stock goes up or down. It would also decide what they want to do with it.